+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 **Node Editor**
 
 - [FP-1516 - Changed the IO/Config Port icons to be more coherent](https://movai.atlassian.net/browse/FP-1516)
+- [FP-1625 - Fixed a few issues regarding Node Editor](https://movai.atlassian.net/browse/FP-1625)
 
 **Home Tab**